requirements.txt
python-telegram-bot==20.3
PyMuPDF==1.23.22
